backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
flake8==7.3.0
mccabe==0.7.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.2.1
sqlparse==0.5.5
